Allow data_items in destination_down()
Although the core DLEP protocol only allows a MAC destination
to be sent in a Destination Down, an extension might want to
send additional data items.  Add a data_items parameter to
DlepService::destination_down() to allow this, and also to
DlepClient::destination_down() to communicate those data items
to the client on the other side.  The new data_items parameter
defaults to an empty list, so calls to the methods won't have
to change.  However, implementations of these methods must
change to conform to the new method signature.