Remove obsolete data item value types
Remove a2_u64, ipv4_u8, ipv6_u8, and u64_u8.  Each of these
was either unused or only used by the ancient draft8, which is
also removed in this commit.  We have to keep the proliferation of
data item value types in check because we use boost::variant to
store the values, and we can currently only have 30 different
types of values.  Culling the unused ones will make room for
new ones.