Fix serialization of u16_vu8 values
There is a count field for the number of u8's in the
vector that wasn't being filled in.  It would have been possible
to serialize this without the count, as the total data item
size is sufficient to tell how many items are in the vector,
but it was defined with a count, so we'll go with that.

Also add a test for this data item value type.