Fail cmake when no C++11 support detected
We check if the compiler supports required C++11 support, but
previously we only wrote out a status message but still continued
the cmake run and added C++ options to the compile.

If, for example, on RHEL/CentOS6, whose default gcc does not have
C++11 support, you could successfully run cmake on these systems,
and then "make" would fail with an error similar to:
error: #error This file requires compiler and library support for
the upcoming ISO C++ standard, C++0x. This support is currently
experimental, and must be enabled with the -std=c++0x or
-std=gnu++0x compiler options.

This commit causes the CMake build to fail when we didn't detect
C++11 support in the compiler.