Set minimum boost version in CMake
Set minimum boost version to 1.48.

RHEL/CentOS 6 comes with 1.41 by default. Dlep.cpp fails with
error: ‘async_connect’ is not a member of ‘boost::asio’

EL6 includes a parallel install of 1.48 (yum install boost148-devel).
This version is acceptable for building DLEP, and this is the oldest
distro we're likely to support, so make this the minimum.

On EL6, hint cmake as to the location of the 1.48 parallel install w/
cmake -DBOOST_INCLUDEDIR=/usr/include/boost148/ -DBOOST_LIBRARYDIR=/usr/lib64/boost148/ ..