fix use-after-free memory error
A boost shared array was being freed before boost::asio async write
was finished with it.  In some environments, this would cause accesses
to freed memory, and garbage to be written by the async write call.

I reworked DlepMessageBuffer to use a shared pointer to vector<uint8_t>,
and to bind the DlepMessageBuffer to the async completion function so
it's sure to stick around until the write is done.

This change will also make it easier to port to the standard c++
library, which doesn't appear to have the equivalent of
boost::shared_array.